+++ test/controllers/sheets_controller_test.rb
@@ -52,6 +52,12 @@ class SheetsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
+  test 'should search for sheets that are on an event' do
+    get :index, params: { project_id: @project, search: 'events:any' }
+    assert_not_nil assigns(:sheets)
+    assert_response :success
+  end
+
   test 'should get index with invalid project' do
     get :index, params: { project_id: -1 }
     assert_nil assigns(:sheets)